chore: API Utils 분리
